Updated the get_ligo_data function
I included a loop that first checks for zeros in a
downsampled version of the data. If it finds zeors
will crop the segments and save the overlaping data
from selected detectors. Now also ignores segments
that data cannot be found. The old version still
exists as old_get_ligo_data.