Fixed bug on auto_far (overestimated data size) and found another one.
After trying auto_far with timeSlides 7 many generation files were
overestimating the available data when they were using the least part
of a segment. I realised that after a set is complete I had to reduce
segment_size_utilisation accordingly. This reduce was of sizes timeSlides^2
in size. Due to that another bug appears, when we use big timeslides the
the reduce of segment_size_utilisation is huge and leaves the file unused.
I need to stop estimating the needed segments at the beggining and do it
as it goes through the loop. The same has to happend tho auto_gen in
datatools.py