HRSS, stackDetector, timeshift and elipticity upgrades
I introduce the option to ask for HRSS instead of SNR. HRSS is a new
known plug in now.

stackDetector can add an extra detector to a dataset on top,
 using the generator function

Experimented with elipticity option on csgs, It will be implemented to more
waveforms in the future

Changed the timeshift of projected injections to be interpolated because
it was affecting the higher frequencies.