Added the gwdatafind server, and reset the default of whitening to weltch
After integrating bonnie's code the default whitening was median.
Until median is the default again, I revert it to welch so that it is
consistent with the trained models.

Added to all condor jobs that use gwdatafind the enviroment variable
os.environ["GWDATAFIND_SERVER"]="datafind.ldas.cit:80"
because it is needed now after a cluster update

Fixed some default accounting group names in validators